tecnomarket/requirements.txt
asgiref==3.4.1
Django==3.0.7
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
pillow
django-admin-interface
django-crispy-forms
django-rest-framework